Convert a vector into text

Writing YAML and Dockerfiles templates requires translating R vectors into strings. In some cases these need to be quoted (e.g. to build R install.packages() commands) and in other cases they need to be space-separated (e.g. Linux yum install a b c).

lam_build_quoted_list(items = NULL, quote = c("double", "single"))

lam_build_separated_list(items = NULL, sep = " ")

Arguments

items
Vector or items to write.

quote
"single" or "double"

Value

A string

Functions

lam_build_quoted_list: Build a double-quoted comma-separated list

lam_build_separated_list: Build an unquoted space-separated list